Refactor storage handling and item management
This commit simplifies and optimizes storage handling and item management. It eliminates a redundant HashMap for 'currentStages', transferring its core functionality to the StorageInventory class. It enhances the efficiency of the system by changing the way current stage of animation is handled and introduces the use of AtomicBoolean to manage the status of storage, improving thread safety. The code also refactors item interface slot determination and reorders the 'save storage' operation in StorageManager class to improve flow control.